% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_estimator.R
\name{DiD}
\alias{DiD}
\title{Estimate DiD for all possible cohorts and event time pairs (g,e), as well as the average across cohorts for each event time (e).}
\usage{
DiD(
  inputdata,
  varname_options,
  control_group = "all",
  event_baseperiod = -1,
  min_event = NULL,
  max_event = NULL
)
}
\arguments{
\item{inputdata}{A data.table.}

\item{varname_options}{A list of the form varname_options = list(id_name, time_name, outcome_name, cohort_name), where all four arguments of the list must be a character that corresponds to a variable name in inputdata.}

\item{control_group}{There are three possibilities: control_group="never-treated" uses the never-treated control group only; control_group="future-treated" uses those units that will receive treatment in the future as the control group; and control_group="all" uses both the never-treated and the future-treated in the control group. Default is control_group="all".}

\item{event_baseperiod}{This is the base pre-period that is normalized to zero in the DiD estimation. Default is event_baseperiod=-1.}

\item{min_event}{This is the minimum event time (e) to estimate. Default is NULL, in which case, no minimum is imposed.}

\item{max_event}{This is the maximum event time (e) to estimate. Default is NULL, in which case, no maximum is imposed.}
}
\value{
A list with two components: results_cohort is a data.table with the DiDge estimates (by event e and cohort g), and results_average is a data.table with the DiDe estimates (by event e, average across cohorts g).
}
\description{
Estimate DiD for all possible cohorts and event time pairs (g,e), as well as the average across cohorts for each event time (e).
}
\examples{
# simulate some data
simdata = SimDiD(sample_size=500)

# define the variable names as a list()
varname_options = list()
varname_options$time_name = "year"
varname_options$outcome_name = "Y"
varname_options$cohort_name = "cohort"
varname_options$id_name = "id"

# estimate the ATT for all cohorts at event times 0-1
DiD(simdata, varname_options, min_event=0, max_event=1)

# change the base period to -3
DiD(simdata, varname_options, event_baseperiod=-3, min_event=0, max_event=1)

# use only the never-treated control group
DiD(simdata, varname_options, control_group = "never-treated", min_event=0, max_event=1)

# use only the future treated control group, check the pre-periods -4 through -2
DiD(simdata, varname_options, control_group = "future-treated", min_event=-4, max_event=-2)
}
