% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_validation.R
\name{EfficiencyTestDiD}
\alias{EfficiencyTestDiD}
\title{Compare speed and memory efficiency of various DiD estimators available in R.}
\usage{
EfficiencyTestDiD(
  estimators = c("DiDforBigData", "CSreg", "CSdr", "CH"),
  sample_sizes = c(1000, 10000, 1e+05),
  reps = 3
)
}
\description{
This function compares speed and memory efficiency of various DiD estimators available in R:
\itemize{
\item \code{DiDforBigData}: This package's DiDforBigData::DiD()
\item \code{saturatedOLS}: This package's fully-interacted OLS (internal)
\item \code{ES}: Novgorodsky and Setzler's Stacked Event-study eventStudy::ES()
\item \code{CSreg}: Callaway and Sant'Anna's reg estimator did::att_gt(..., est_method = "reg")
\item \code{CSdr}: Callaway and Sant'Anna's dr estimator did::att_gt(..., est_method = "dr")
\item \code{CH}: DIDmultiplegt::did_multiplegt()
}
}
