% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_plots.R
\name{PlotES}
\alias{PlotES}
\title{Plot an event study in the original units.}
\usage{
PlotES(
  DiD_results,
  plot_type,
  cohort = NULL,
  ci_factor = 1.96,
  max_jitter = 0.05,
  lower_event = -3,
  upper_event = 5,
  colors = NULL
)
}
\arguments{
\item{DiD_results}{Output from the DiD function.}

\item{plot_type}{There are 2 options: default option is plot_type="ge" plots all the event time (e) for a given cohort (g); and option plot_type="e" plots the average across cohorts for each event time (e).}

\item{cohort}{Select one cohort to plot if plot_type="ge".}

\item{ci_factor}{This is the multiplier that converts a standard errors (SE) into a confidence interval (CI). That is, CI = estimate +/- ci_factor*SE. Default option is ci_factor = 1.96, which corresponds to a 95\% CI. A natural alternative is ci_factor = 1.645, which corresponds to a 90\% CI.}

\item{max_jitter}{This jitters the points so that they do not overlap. If plot_type="ge" or plot_type="gt", there will be multiple points overlapping due to multiple cohorts, and max_jitter>0 will dodge the points by at most max_jitter. Default option is max_jitter=0.05.}

\item{lower_event}{Minimum event time (e) to show in the plot. Default option is NULL.}

\item{upper_event}{Maximum event time (e) to show in the plot. Default option is NULL.}

\item{colors}{Manually set the colors for each cohort. Only relevant if plot_type="ge" or plot_type="gt". Default option is NULL.}
}
\value{
A ggplot object.
}
\description{
Plot an event study in the original units.
}
