% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_estimator.R
\name{DiD}
\alias{DiD}
\title{Estimate DiD for all possible cohorts and event time pairs (g,e), as well as the average across cohorts for each event time (e).}
\usage{
DiD(
  inputdata,
  varnames,
  control_group = "all",
  baseperiod = -1,
  min_event = NULL,
  max_event = NULL,
  Esets = NULL
)
}
\arguments{
\item{inputdata}{A data.table.}

\item{varnames}{A list of the form varnames = list(id_name, time_name, outcome_name, cohort_name), where all four arguments of the list must be a character that corresponds to a variable name in inputdata.}

\item{control_group}{There are three possibilities: control_group="never-treated" uses the never-treated control group only; control_group="future-treated" uses those units that will receive treatment in the future as the control group; and control_group="all" uses both the never-treated and the future-treated in the control group. Default is control_group="all".}

\item{baseperiod}{This is the base pre-period that is normalized to zero in the DiD estimation. Default is baseperiod=-1.}

\item{min_event}{This is the minimum event time (e) to estimate. Default is NULL, in which case, no minimum is imposed.}

\item{max_event}{This is the maximum event time (e) to estimate. Default is NULL, in which case, no maximum is imposed.}

\item{Esets}{If a list of sets of event times is provided, it will loop over those sets, computing the average ATT_e across event times e. Default is NULL.}
}
\value{
A list with two components: results_cohort is a data.table with the DiDge estimates (by event e and cohort g), and results_average is a data.table with the DiDe estimates (by event e, average across cohorts g).
}
\description{
Estimate DiD for all possible cohorts and event time pairs (g,e), as well as the average across cohorts for each event time (e).
}
\examples{
# simulate some data
simdata = SimDiD(sample_size=1000, ATTcohortdiff = 2)$simdata

# define the variable names as a list()
varnames = list()
varnames$time_name = "year"
varnames$outcome_name = "Y"
varnames$cohort_name = "cohort"
varnames$id_name = "id"

# estimate the ATT for all cohorts at event time 1 only
DiD(simdata, varnames, min_event=1, max_event=1)

# change the base period to -3
DiD(simdata, varnames, baseperiod=-3, min_event=1, max_event=1)

# check the pre-periods -4 through -2
DiD(simdata, varnames, control_group = "never-treated", min_event=-4, max_event=-2)

# use only the never-treated control group, estimate events -4 through 1
DiD(simdata, varnames, control_group = "never-treated", min_event=-4, max_event=1)

# use only the future treated control group, estimate events -4 through 1
DiD(simdata, varnames, control_group = "future-treated", min_event=-4, max_event=1)

# estimate average ATTe across sets of events
DiD(simdata, varnames, min_event=-4, max_event=6, Esets=list(c(-3,-2,-1),c(1,2,3)))
}
