% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_estimator.R
\name{DiDge}
\alias{DiDge}
\title{Estimate DiD for a single cohort (g) and a single event time (e).}
\usage{
DiDge(
  inputdata,
  varnames,
  cohort_time,
  event_postperiod,
  baseperiod = -1,
  control_group = "all"
)
}
\arguments{
\item{inputdata}{A data.table.}

\item{varnames}{A list of the form varnames = list(id_name, time_name, outcome_name, cohort_name), where all four arguments of the list must be a character that corresponds to a variable name in inputdata.}

\item{baseperiod}{This is the base pre-period that is normalized to zero in the DiD estimation. Default is baseperiod=-1.}

\item{control_group}{There are three possibilities: control_group="never-treated" uses the never-treated control group only; control_group="future-treated" uses those units that will receive treatment in the future as the control group; and control_group="all" uses both the never-treated and the future-treated in the control group. Default is control_group="all".}

\item{min_event}{This is the minimum event time (e) to estimate. Default is NULL, in which case, no minimum is imposed.}

\item{max_event}{This is the maximum event time (e) to estimate. Default is NULL, in which case, no maximum is imposed.}
}
\value{
A list with two components: results_cohort is a data.table with the DiDge estimates (by event e and cohort g), and results_average is a data.table with the DiDe estimates (by event e, average across cohorts g).
}
\description{
Estimate DiD for a single cohort (g) and a single event time (e).
}
\examples{
# simulate some data
simdata = SimDiD(sample_size=200)

# define the variable names as a list()
varnames = list()
varnames$time_name = "year"
varnames$outcome_name = "Y"
varnames$cohort_name = "cohort"
varnames$id_name = "id"

# estimate the ATT for cohort 2007 at event time 1
DiDge(simdata, varnames, cohort_time=2007, event_postperiod=1)

# change the base period to -3
DiDge(simdata, varnames, baseperiod=-3, cohort_time=2007, event_postperiod=1)

# use only the never-treated control group
DiDge(simdata, varnames, control_group = "never-treated", cohort_time=2007, event_postperiod=1)

# use only the never-treated control group
DiDge(simdata, varnames, control_group = "future-treated", cohort_time=2007, event_postperiod=1)
}
