% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiD_simulator.R
\name{SimDiD}
\alias{SimDiD}
\title{DiD data simulator with staggered treatment.}
\usage{
SimDiD(
  seed = 1,
  sample_size = 100,
  cohorts = c(2007, 2010, 2012),
  ATTs = c(1, 1, 1),
  anticipation = 0,
  minyear = 2003,
  maxyear = 2013,
  idvar = 1,
  yearvar = 1,
  shockvar = 1
)
}
\arguments{
\item{seed}{Set the random seed. Default is seed=1.}

\item{sample_size}{Number of individuals. Default is sample_size=100.}

\item{cohorts}{Vector of years at which treatment onset occurs. Default is cohorts=c(2007,2010,2012).}

\item{ATTs}{Vector of treatment effects. Must be same length as cohorts. Default is ATTs=c(1,1,1).}

\item{anticipation}{Number of years prior to cohort to allow 50\% treatment effects. Default is anticipation=0.}

\item{minyear}{Minimum calendar year to include in the data. Default is minyear=2003.}

\item{maxyear}{Maximum calendar year to include in the data. Default is maxyear=2013.}

\item{idvar}{Variance of individual fixed effects (alpha_i). Default is idvar=1.}

\item{yearvar}{Variance of year effects (mu_i). Default is yearvar=1.}

\item{shockvar}{Variance of idiosyncratic shocks (epsilon_it). Default is shockvar=1.}
}
\value{
A data.table with variables (id, year, cohort, Y), where Y is the outcome variable.
}
\description{
Simulate data from the model Y_it =  alpha_i + mu_t + ATT*(t >= G_i) + epsilon_it, where i is individual, t is year, and G_i is the cohort.
}
\examples{
SimDiD()
}
